+++ CHANGELOG.md
@@ -11,6 +11,7 @@ to a section with the version name.
 * Support multiple billing references for return invoices
   * Add `Return Against Additional References` custom field to sales invoice
   * The field has no ERPNext impact. The additional references are included in the XML for ZATCA
+  * The field allows submitted non-return invoices for the same company/customer/supplier as the current invoice
 
 ## 0.43.2